bugfix: add safer function to find indices than dsearchn
Previously, we have been using dsearchn to quickly find indices of
lags/frequencies of interest. However, dsearchn always returns, finding
the element that is **closest** to the requested lag or frequency. This
can result in problematic behavior. For example, imagine we have a
vector of frequencies only up to 20 Hz, but we request to find
meter-related frequency indices up to 30 Hz. With dsearchn, all the
meter-related frequencies above 20 Hz will end up with the same index,
corresponding to the highest frequency in the array, i.e. 20 Hz.

To fix this, I have written a custom function that does several checks.

I'm also adding some unit tests for it.