defaults for find_idx tolerance
After adding the find_idx_tol.m function, there is a tolerance
threshold for finding indices of lags/freq of interest.

If there is no value in the array of frequencies/lags that is closer
than `tol` to the requested frequency/lag we want to find index for,
this value will be ignored.

However, sometimes we have a sampling rate that doesn't allow for
integer number of samples per trial, and so we will not end up with
frequncy bins exactly at the frequencies of interest. Same goes for
autocorrelation lags.

Hence, I tried to add reasonable defaults to the get_acf_features and
get_fft_features function. The parameter is passed onto find_idx_tol.