set acf normalization true by default (lag 0 = 1)
We will set the normalization option to be true by default. This means
that the output acf values will be equivalent to taking Pearson's
correlations with the input signal x and it's time-shifted versions.

The only thing to make this happen is to normalize the resulting ACF by
the value at lag 0, making the lag0 value equal to 1.