bugfix: get f0 harm outside of rm_ap section
Before, the F0 harmonics and their indices were obtained inside the
section where 1/f is removed. The section that only keeps F0 and
harmonics would be only run when these variables were in the workspace,
so it was silently skipped when the user asked to keep F0 harm without
1/f subtraction.

Now we get the F0 harmonics at the top of the script, since they are
required in several sections, incl. 1/f removal with fooof, and when
only keeping F0 harmonics.